+++ app/PersonioIntegration/Imports/Api.php
@@ -21,6 +21,7 @@ use PersonioIntegrationLight\PersonioIntegration\Imports_Base;
 use PersonioIntegrationLight\PersonioIntegration\Personio;
 use PersonioIntegrationLight\PersonioIntegration\Position;
 use PersonioIntegrationLight\PersonioIntegration\Positions;
+use PersonioIntegrationLight\Plugin\Schedules\ApiAccessToken;
 use stdClass;
 use WP_Post;
 
@@ -97,7 +98,7 @@ class Api extends Imports_Base {
 	 */
 	public function run(): void {
 		// if debug mode is enabled, log this event.
-		Log::get_instance()->add( __( 'Import of positions is now running.', 'personio-integration-light' ), 'info', 'import' );
+		Log::get_instance()->add( __( 'Trying to authenticate on the Personio API.', 'personio-integration-light' ), 'info', 'import' );
 
 		// get the API object.
 		$api_obj = \PersonioIntegrationLight\PersonioIntegration\Api::get_instance();
@@ -108,7 +109,7 @@ class Api extends Imports_Base {
 		// bail if no token is set.
 		if ( empty( $access_token ) ) {
 			// log this as an error.
-			$this->add_error( __( 'No Access token for API available. Import from API will not run.', 'personio-integration-light' ) );
+			$this->add_error( __( 'No access token for API available. Import from API will not run.', 'personio-integration-light' ) );
 
 			// reset status.
 			update_option( WP_PERSONIO_INTEGRATION_IMPORT_STATUS, '' );
@@ -116,6 +117,9 @@ class Api extends Imports_Base {
 			// do nothing more.
 			return;
 		}
+
+		// if debug mode is enabled, log this event.
+		Log::get_instance()->add( __( 'Import of positions is now running.', 'personio-integration-light' ), 'info', 'import' );
 
 		// set the mark that import is running in WP.
 		if ( ! defined( 'WP_IMPORTING' ) ) {
@@ -558,5 +562,37 @@ class Api extends Imports_Base {
 	 */
 	public function can_be_enabled_by_user(): bool {
 		return Helper::is_development_mode_active();
+	}
+
+	/**
+	 * Toggle the state of this extension and manage its schedule accordingly.
+	 *
+	 * @return void
+	 */
+	public function toggle_state(): void {
+		parent::toggle_state();
+
+		// get the schedule object.
+		$schedule_obj = new ApiAccessToken();
+
+		// get the actual state of this extension.
+		$state = absint( get_option( $this->get_settings_field_name() ) );
+
+		if ( 1 === $state ) {
+			// install the schedule if credentials are set.
+			if( \PersonioIntegrationLight\PersonioIntegration\Api::get_instance()->is_credential_prepared() ) {
+				update_option( 'personioIntegrationEnableApiAccessToken', 1 );
+				$schedule_obj->install();
+			}
+
+			// do nothing more.
+			return;
+		}
+
+		// extension has been disabled: remove the schedule and the token, independent of the secret.
+		$schedule_obj->delete();
+
+		// set token to "disabled".
+		update_option( 'personioIntegrationEnableApiAccessToken', 0 );
 	}
 }