+++ changelog.md
@@ -2,9 +2,19 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added Personio icon in the WordPress icon library (for WordPress >= 7.1)
+- Added new logging events during plugin activation and deactivation
+
+### Changed
+
+- Small updates for die API v2, but still not usable as Personio do not deliver the most of v1 data for positions
+
 ### Fixed
 
 - Fixed the statement to get our own Block templates from the database (thanks @philipp-marquardt)
+- Fixed removing of schedules with attributes during deactivation of plugin
 
 ## [5.6.0] - 18.08.2026